refactor(#3096): split the WIRE-side FlightData framing governor out of batch_bytes
The review fix pushed `batch_bytes.rs` from 650 to 898 lines, past the 800-line
campsite threshold, and the lite gate's file-size ratchet FAILed it. Split by
responsibility rather than waived with CQLITE_ALLOW_FILE_GROWTH:

- NEW `cqlite-flight/src/flight_data_size.rs` (485 lines) — the wire-side
  CAPACITY governor: GRPC_DEFAULT_MAX_MESSAGE_BYTES with its primary sources, the
  framing reserve / inexactness margin / reserved ceiling, the derived shipped
  target, the superseded + rejected targets with their MEASURED_* figures, and
  every compile-time guard.
- `batch_bytes.rs` (449 lines) keeps the producer-side PAYLOAD cap: the
  --max-batch-bytes contract, BatchByteCap, split_rows_into_batches and the
  capacity conversion. Its module doc now points next door.

This is the boundary the defect itself argued for: conflating a producer PAYLOAD
cap with a wire CAPACITY target is exactly how the encoder target came to sit ON
the gRPC ceiling. The two governors are now separate modules, so the distinction
is structural instead of a paragraph.

Pure move plus import/doc-link repointing (streaming.rs, streaming_framing_tests.rs,
lib.rs); no constant, guard or assertion changed value. All 6 framing tests and 23
batch_bytes tests stay green; the file-size ratchet now reports no grown
over-threshold file.
